+++ src/Scopus/Response/Affiliation.php
@@ -40,14 +40,14 @@ class Affiliation
         return isset($this->data['affiliation-country']) ? $this->data['affiliation-country'] : null;
     }
     
-    public function getNameVariant()
+    public function getNameVariant(): array
     {
         if (isset($this->data['name-variant'])) {
-            return $this->nameVariants ?: $this->nameVariants = array_map(function($nameVariant) {
+            return $this->nameVariants ?? $this->nameVariants = array_map(function($nameVariant) {
                 return $nameVariant['$'];
             }, $this->data['name-variant']);
         }
 
-        return null;
+        return [];
     }
 }