+++ src/Scopus/Response/BibrecordHead.php
@@ -24,7 +24,7 @@ class BibrecordHead
     public function getAuthorGroup()
     {
         if (isset($this->data['author-group'])) {
-            return $this->authorGroup ?: $this->authorGroup = new AuthorGroup($this->data['author-group']);
+            return $this->authorGroup ?? $this->authorGroup = new AuthorGroup($this->data['author-group']);
         }
 
         return null;
@@ -33,7 +33,7 @@ class BibrecordHead
     public function getCorrespondence()
     {
         if (isset($this->data['correspondence'])) {
-            return $this->correspondence ?: $this->correspondence = new Correspondence($this->data['correspondence']);
+            return $this->correspondence ?? $this->correspondence = new Correspondence($this->data['correspondence']);
         }
 
         return null;
@@ -42,7 +42,7 @@ class BibrecordHead
     public function getSource()
     {
         if (isset($this->data['source'])) {
-            return $this->source ?: $this->source = new Source($this->data['source']);
+            return $this->source ?? $this->source = new Source($this->data['source']);
         }
 
         return null;